+++ src/game/proto.h
@@ -6,9 +6,9 @@
 
 bool proto_init(GameInitInfo* init_info);
 void proto_exit();
-long long sub_468570(ObjectType object_type);
-long long sub_4685A0(int a1);
+int64_t sub_468570(ObjectType object_type);
+int64_t sub_4685A0(int a1);
 bool sub_468610(int description);
-int sub_49B290(long long obj);
+int sub_49B290(int64_t obj);
 
 #endif /* ARCANUM_GAME_PROTO_H_ */
