+++ src/ui/tb_ui.h
@@ -6,7 +6,7 @@
 bool tb_ui_init(GameInitInfo* init_info);
 void tb_ui_reset();
 void tb_ui_exit();
-void sub_57CCF0(long long a1, long long a2);
-bool sub_57CD60(long long a1, long long a2, char* buffer);
+void sub_57CCF0(int64_t a1, int64_t a2);
+bool sub_57CD60(int64_t a1, int64_t a2, char* buffer);
 
 #endif /* ARCANUM_UI_TB_UI_H_ */
